+++ src/new-components/button-link/button-link.test.tsx
@@ -96,7 +96,7 @@ describe('ButtonLink', () => {
         )
         const buttonLink = screen.getByRole('link', { name: 'Click me now' })
         expect(buttonLink.innerHTML).toMatchInlineSnapshot(
-            `"<span data-testid=\\"button-nested-span\\" class=\\"label box textAlign-center overflow-hidden\\">Click me <strong>now</strong></span>"`,
+            `"<span class=\\"label box textAlign-center overflow-hidden\\">Click me <strong>now</strong></span>"`,
         )
     })
 
@@ -238,24 +238,6 @@ describe('ButtonLink', () => {
         )
         const buttonLink = screen.getByRole('link', { name: 'Click me' })
         expect(buttonLink).not.toHaveClass('align-end')
-    })
-
-    it('contains a non-clickable nested span', () => {
-        const onClick = jest.fn()
-        render(
-            <ButtonLink href="/" variant="primary">
-                Click me
-            </ButtonLink>,
-        )
-        const nestedSpan = screen.getByTestId('button-nested-span')
-        // Verify that the button contains a nested span
-        expect(screen.getByRole('link', { name: 'Click me' })).toContainElement(nestedSpan)
-        // Set an onClick function on the span
-        nestedSpan.onclick = onClick
-        // Click on the button
-        userEvent.click(nestedSpan)
-        // Ensure that the onClick function set on the span has not been called, meaning that the span is non-clickable
-        expect(onClick).not.toHaveBeenCalled()
     })
 
     describe('with icons', () => {